added two basic pipeline examples to a new pipelines.examples module. updated tests to import those rather than (re-)define
