+++ sdk/bettmensch_ai/pipelines/examples/__init__.py
@@ -1,0 +1,4 @@
+from bettmensch_ai.pipelines.examples.basic import (  # noqa: F401
+    adding_parameters_pipeline,
+    parameter_to_artifact_pipeline,
+)
